+++ symphony-bdk-core/src/main/java/com/symphony/bdk/core/activity/parsing/input/Hashtag.java
@@ -1,16 +1,16 @@
-package com.symphony.bdk.core.activity.parsing;
+package com.symphony.bdk.core.activity.parsing.input;
 
 import lombok.Getter;
 import org.apiguardian.api.API;
 
 @API(status = API.Status.INTERNAL)
 @Getter
-public class Cashtag {
+public class Hashtag {
 
   private String text;
   private String value;
 
-  public Cashtag(String text, String value) {
+  public Hashtag(String text, String value) {
     this.text = text;
     this.value = value;
   }
