+++ symphony-bdk-core/src/main/java/com/symphony/bdk/core/activity/parsing/input/InputTokenizer.java
@@ -1,4 +1,4 @@
-package com.symphony.bdk.core.activity.parsing;
+package com.symphony.bdk.core.activity.parsing.input;
 
 import static org.apache.commons.lang3.StringUtils.isBlank;
